+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fix a problem confining ICU version if it's located in a different
   directory. See `Wrapper.SetPreferredIcu4cDirectory()`.
 - Include `icu.net.dll.config` file in nuget package. This is important for running on MacOSX.
+- Fix construction of locale with language and keywords (cbersch)
 
 ## [2.7.1] - 2021-03-04