+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Include `icu.net.dll.config` file in nuget package. This is important for running on MacOSX.
 - Fix construction of locale with language and keywords (cbersch)
 - Fix passing locale to ubrk_open (cbersch)
+- Fix race condition during initialization of native methods container (cbersch)
 
 ## [2.7.1] - 2021-03-04